removed simple-crypt and cleaned up usage of keys for encryption w/ burnettk
