+++ spiffworkflow-backend/src/spiffworkflow_backend/config/unit_testing.py
@@ -3,7 +3,6 @@ import os
 from os import environ
 
 TESTING = True
-SECRET_KEY = "the_secret_key"  # noqa: S105, do not care about security when running unit tests
 SPIFFWORKFLOW_BACKEND_LOG_TO_FILE = environ.get("SPIFFWORKFLOW_BACKEND_LOG_TO_FILE", default="true") == "true"
 
 SPIFFWORKFLOW_BACKEND_PERMISSIONS_FILE_NAME = environ.get(
